feat: Implement `movetoPos` in Spindex
Adds a `movetoPos` method to the `Spindex` subsystem. This method moves the spindex motor to a target angle based on the voltage reading from an analog sensor. It uses a proportional control loop to slow down as it approaches the target position.

Also includes minor whitespace changes in the `Test.java` file.